davinci: switch back to AIDL power hal
Revert "power-libperfmgr: create stune group for camera, drop unsupported hints"

This reverts commit 2b3ed5b749371e5d3db4b46f5d0c18a9b3485e9c.

Revert "power-libperfmgr: remove Google-specific display LPM control"

This reverts commit b0ea8e2c7eca7e4d1ec7366bd12dac4257d54ee4.

Revert "power-libperfmgr: Clean up device layout"

This reverts commit ba8001e533607d4b56f9eef30a7788991db0fe6d.

Revert "power-libperfmgr: Removing support for stats related APIs"

This reverts commit e35d243c5bf6cac95b7b6ccc487f031ffe4e287f.

Revert "power-libperfmgr: Rewrite display-helper into cpp"

This reverts commit 065ece81e6701775a0a9253b97d6e9d7d8e1e018.

Revert "davinci: start the Power HAL after writing the scheduler values"

This reverts commit 2655cb3e07ad0da8d8495c02cda88ea99e440808.

Revert "davinci: power-libperfmgr: Scan input devices for eligible DT2W node"

This reverts commit c632bd65cbecc049545af55b4b7cae6d6d7199cc.

Revert "davinci: Add support for event-based tap to wake gesture control"

This reverts commit 560537bf33e4ad8e461772be4d476784b1b7d940.

Revert "davinci: sepolicy: Add libperfmgr rules"

This reverts commit 44c6753c772b1e3e599884dfedab1ea611ee9d14.

Revert "davinci: import Pixel 4a powerhint"

This reverts commit 1327a7fb11e6b4e4a002b1d548bfd8630e8f1c1e.

Revert "davinci: add android.hardware.power to the vintf manifest"

This reverts commit 1bc201a0642b871f20d4724165dc0336417d5dc9.

Revert "davinci: include hardware/google/pixel as a soong namespace"

This reverts commit c49bc2dc8760d5f92b669154ceb8268b75804140.

Revert "davinci: remove unneeded perf blobs"

This reverts commit 55df2112c44711371ab23967390f4c179a67e831.

Revert "davinci: power-libperfmgr: change WLAN stats node"

This reverts commit be8e86bf918a6d90ce0ab46b4dac9ecfb53c2f85.

Revert "davinci: switch to libperfmgr for powerhint"

This reverts commit cf44b1003b39118030a69f319b072cb24fa98bfd.
