+++ Android.bp
@@ -1,6 +1,5 @@
 soong_namespace {
     imports: [
         "hardware/xiaomi",
-        "hardware/google/pixel"
     ],
 }
