+++ device.mk
@@ -300,8 +300,7 @@ PRODUCT_ENFORCE_RRO_EXCLUDED_OVERLAYS += \
 
 # Power
 PRODUCT_PACKAGES += \
-    android.hardware.power@1.3-service.sm6150-libperfmgr \
-    powerhint.json
+    android.hardware.power-service-qti
 
 # Prebuilts
 PRODUCT_PACKAGES += \
@@ -354,8 +353,7 @@ PRODUCT_COPY_FILES += \
 # Soong namespaces
 PRODUCT_SOONG_NAMESPACES += \
     $(LOCAL_PATH) \
-    hardware/xiaomi \
-    hardware/google/pixel
+    hardware/xiaomi
 
 # Telephony
 PRODUCT_PACKAGES += \
