--- power-libperfmgr/Android.bp
@@ -1,46 +1,0 @@
-//
-// Copyright (C) 2018 The Android Open Source Project
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-cc_binary {
-    name: "android.hardware.power@1.3-service.sm6150-libperfmgr",
-    relative_install_path: "hw",
-    init_rc: ["android.hardware.power@1.3-service.sm6150-libperfmgr.rc"],
-        srcs: [
-        "service.cpp",
-        "Power.cpp",
-        "InteractionHandler.cpp",
-    ],
-    cflags: [
-        "-Wall",
-        "-Werror",
-    ],
-    shared_libs: [
-        "libbase",
-        "libhidlbase",
-        "liblog",
-        "libutils",
-        "libcutils",
-        "android.hardware.power@1.0",
-        "android.hardware.power@1.1",
-        "android.hardware.power@1.2",
-        "android.hardware.power@1.3",
-        "libperfmgr",
-    ],
-	header_libs: [
-        "libhardware_headers",
-        "generated_kernel_headers",
-    ],
-    required: ["powerhint.json"],
-    proprietary: true,
-}
